Data.MemoCombinators.Class

Documentation

class MemoTable a where

The class of types which have complete memo tables.

Minimal complete definition

table

Methods

table :: Memo a

class Memoizable a where

The class of functions which can be completely memoized.

Minimal complete definition

memoize

Methods

memoize :: a -> a