Data.MemoCombinators.Class

Documentation

class MemoTable a where

The class of types which have complete memo tables.

Methods

table :: Memo a

class Memoizable a where

The class of functions which can be completely memoized.

Methods

memoize :: a -> a